+++ src/main/scala/sc4pac/api/message.scala
@@ -303,7 +303,7 @@ case class ChannelStatsAll(combined: JD.Channel.Stats, channels: Seq[ChannelStat
 
 case class InitArgs(plugins: String, cache: String, temp: String) derives UP.ReadWriter
 
-case class ServerStatus(sc4pacVersion: String) derives UP.ReadWriter
+case class ServerStatus(sc4pacVersion: String, osVersion: String) derives UP.ReadWriter
 
 case class ProfileName(name: String) derives UP.ReadWriter
 case class ProfileIdObj(id: ProfileId) derives UP.ReadWriter