+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - API upgraded to 2.6:
   - added `variantChoices` to `/packages.info` output
   - added new `/variants.choices` and `/variants.set` endpoints
+  - added `osVersion` to `/server.status` endpoint
 
 
 ## [0.7.0] - 2025-05-11